refactor: reorganize DTO packages and update imports for consistency
