--- src/main/java/br/com/rodrigo/onlinelibraryapi/dtos/TokenJWT.java
@@ -1,5 +1,0 @@
-package br.com.rodrigo.onlinelibraryapi.dtos;
-
-public record TokenJWT(String token) {
-    
-}
