+++ src/main/java/br/com/rodrigo/onlinelibraryapi/dtos/author/ListAuthorDTO.java
@@ -1,0 +1,5 @@
+package br.com.rodrigo.onlinelibraryapi.dtos.author;
+
+public record ListAuthorDTO() {
+    
+}
