+++ src/main/java/br/com/rodrigo/onlinelibraryapi/dtos/token/TokenJWT.java
@@ -1,0 +1,5 @@
+package br.com.rodrigo.onlinelibraryapi.dtos.token;
+
+public record TokenJWT(String token) {
+    
+}
