+++ src/main/java/br/com/rodrigo/onlinelibraryapi/entities/User.java
@@ -6,7 +6,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import br.com.rodrigo.onlinelibraryapi.dtos.CreateUserDto;
+import br.com.rodrigo.onlinelibraryapi.dtos.user.CreateUserDto;
 import br.com.rodrigo.onlinelibraryapi.entities.embedded.Address;
 import br.com.rodrigo.onlinelibraryapi.entities.embedded.Authentication;
 import br.com.rodrigo.onlinelibraryapi.entities.embedded.Name;
